Mount v2 of the application
